Guess the number
This program generates a random number from 0 to 20 and you have 7 chances to guess the number(as 7 is my favorite number). After every guess the program tells you if the auto generated number is greater or smaller than your guessed number and the number guesses left. If you guess the it tells you that you have guessed the right number and exactly how many guesses you took to guess it right. If you run out of guesses it tells you that you lost and what the number actually was.